add constants add module style fix module styles bug fix key warning with nanoid
