+++ package.json
@@ -4,6 +4,7 @@
   "description": "",
   "scripts": {
     "start": "webpack serve --mode=development",
+    "build": "webpack --mode production",
     "lint": "npx eslint ./ --ext .js,.jsx,.ts,.tsx --fix",
     "prepare": "husky install"
   },
@@ -11,54 +12,55 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@babel/core": "^7.17.2",
+    "@babel/core": "^7.17.5",
     "@babel/node": "^7.16.8",
     "@babel/preset-env": "^7.16.11",
     "@babel/preset-react": "^7.16.7",
     "@babel/preset-typescript": "^7.16.7",
-    "@storybook/addon-actions": "^6.4.18",
-    "@storybook/addon-essentials": "^6.1.21",
-    "@storybook/addon-links": "^6.4.18",
-    "@storybook/react": "^6.1.21",
+    "@storybook/addon-actions": "^6.4.19",
+    "@storybook/addon-essentials": "^6.4.19",
+    "@storybook/addon-links": "^6.4.19",
+    "@storybook/react": "^6.4.19",
     "@types/react-dom": "^17.0.11",
-    "@typescript-eslint/eslint-plugin": "^5.12.0",
+    "@typescript-eslint/eslint-plugin": "^5.12.1",
     "@webpack-cli/serve": "^1.6.1",
     "babel-loader": "^8.2.3",
-    "eslint": "^8.8.0",
-    "eslint-config-prettier": "^8.3.0",
+    "eslint": "^8.9.0",
+    "eslint-config-prettier": "^8.4.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.28.0",
+    "eslint-plugin-react": "^7.29.0",
     "eslint-plugin-react-hooks": "^4.3.0",
-    "eslint-plugin-storybook": "^0.5.6",
+    "eslint-plugin-storybook": "^0.5.7",
     "husky": "^7.0.4",
     "loki": "^0.28.1",
     "prettier": "^2.5.1",
     "ts-loader": "^9.2.6",
     "typescript": "^4.5.5",
-    "webpack": "^5.68.0",
+    "webpack": "^5.69.1",
     "webpack-cli": "^4.9.2",
     "webpack-dev-middleware": "^5.3.1",
     "webpack-dev-server": "^4.7.4",
     "webpack-hot-middleware": "^2.25.1"
   },
   "dependencies": {
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
+    "@emotion/react": "^11.8.1",
+    "@emotion/styled": "^11.8.1",
     "@material-ui/core": "^4.12.3",
     "@material-ui/icons": "^4.11.2",
-    "@mui/icons-material": "^5.4.1",
-    "@mui/material": "^5.4.1",
-    "@mui/styled-engine-sc": "^5.4.1",
-    "@typescript-eslint/parser": "^5.12.0",
+    "@mui/icons-material": "^5.4.2",
+    "@mui/material": "^5.4.3",
+    "@mui/styled-engine-sc": "^5.4.2",
+    "@typescript-eslint/parser": "^5.12.1",
     "css-loader": "^6.6.0",
     "html-webpack-plugin": "^5.5.0",
-    "nanoid": "^3.2.0",
+    "nanoid": "^3.3.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "^5.0.0",
     "style-loader": "^3.3.1",
     "styled-components": "^5.3.3",
-    "svg-url-loader": "^7.1.1"
+    "svg-url-loader": "^7.1.1",
+    "typescript-plugin-css-modules": "^3.4.0"
   },
   "loki": {
     "configurations": {
