+++ src/App.module.css
@@ -9,6 +9,9 @@ html {
   margin-right: auto;
   padding-left: 16px;
   padding-right: 16px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
 }
 .messageWrap {
   display: flex;
