+++ src/App.tsx
@@ -1,52 +1,10 @@
 import React, { useState, useEffect, FC } from 'react';
-import './App.css';
+import style from './App.module.css';
 import { nanoid } from 'nanoid';
-import { MessageList } from './components/MessageList/MessageList';
+import { Message, MessageList } from './components/MessageList/MessageList';
 import { Form } from './components/Form/Form';
-
-export interface Message {
-  id: string;
-  text: string;
-  author: string;
-}
-
-const botAuthors: string[] = [
-  'Jacque Fresco',
-  'Philippe Starck',
-  'Victor J. Papanek',
-  'Dieter Rams',
-  'Thomas Chippendale',
-  'Karim Rashid',
-  'Ron Arad',
-];
-
-const botText: string[] = [
-  'Лучше быть последним — первым, чем первым — последним.',
-  'На случай, если буду нужен, то я там же, где и был, когда был не нужен.',
-  'Если волк молчит то лучше его не перебивай.',
-  'Каждый в цирке думает, что знает в цирке, но не каждый, что в цирке знает, что в цирке не каждый знает думает.',
-  'Легко вставать, когда ты не ложился.',
-  'За двумя зайцами погонишься — рыбку из пруда не выловишь, делу время, а отмеришь семь раз.',
-  'Каждый думает, что не знает что, но не каждый не знает, что знает, кто не я...',
-  'Кем бы ты ни был, кем бы ты не стал, помни, где ты был и кем ты стал.',
-];
-
-function arrRandom(arr: string[]) {
-  const rand = Math.floor(Math.random() * arr.length);
-  return arr[rand];
-}
-
-interface AppState {
-  messages: Message[];
-}
-
-const defaultMessages = [
-  {
-    id: '1',
-    author: 'Jacque Fresco',
-    text: 'Welcome to the chat',
-  },
-];
+import { defaultMessages, botAuthors, botText } from './constants/Constants';
+import { arrRandom } from './utils/arrRandom';
 
 export const App: FC = () => {
   const [messages, setMessages] = useState<Message[]>(defaultMessages);
@@ -79,8 +37,10 @@ export const App: FC = () => {
   };
 
   return (
-    <div className="messageWrap container">
-      <h1 className="messageTitle">Welcome to react</h1>
+    <div className={(style.messageWrap, style.container)}>
+      <h1 className={style.messageTitle} key={nanoid()}>
+        Welcome to react
+      </h1>
       <MessageList messages={messages} />
       <Form addMessage={addMessage} />
     </div>
