+++ src/components/Form/Form.tsx
@@ -1,8 +1,7 @@
 import React, { useState, FC } from 'react';
 import { Input, Button } from '@mui/material';
 import SendRoundedIcon from '@material-ui/icons/SendRounded';
-import PropTypes from 'prop-types';
-import './Form.css';
+import style from './Form.module.css';
 
 interface Message {
   text: string;
@@ -27,14 +26,10 @@ export const Form: FC<FormProps> = ({ addMessage }) => {
 
   return (
     <form onSubmit={handleText}>
-      <Input className="formInput" value={text} onChange={(ev) => setText(ev.target.value)} />
+      <Input className={style.formInput} value={text} onChange={(ev) => setText(ev.target.value)} />
       <Button variant="contained" startIcon={<SendRoundedIcon />} type="submit">
         Send
       </Button>
     </form>
   );
-};
-
-Form.propTypes = {
-  addMessage: PropTypes.any,
 };
