+++ src/components/MessageList/MessageList.tsx
@@ -1,16 +1,20 @@
 import React, { FC } from 'react';
-import { Message } from '../../App';
-import './Message.css';
+import style from './Message.module.css';
 
+export interface Message {
+  id: string;
+  text: string;
+  author: string;
+}
 interface MessageListProp {
   messages: Message[];
 }
 
 export const MessageList: FC<MessageListProp> = ({ messages }) => (
-  <ul className="messageList">
+  <ul className={style.messageList}>
     {messages.map((message) => (
-      <li className="message" key={message.id}>
-        <span className="messageAuthor">{message.author}:</span> {message.text}
+      <li className={style.message} key={message.id}>
+        <span className={style.messageAuthor}>{message.author}:</span> {message.text}
       </li>
     ))}
   </ul>
