+++ src/constants/Constants.tsx
@@ -1,0 +1,28 @@
+export const botAuthors: string[] = [
+  'Jacque Fresco',
+  'Philippe Starck',
+  'Victor J. Papanek',
+  'Dieter Rams',
+  'Thomas Chippendale',
+  'Karim Rashid',
+  'Ron Arad',
+];
+
+export const botText: string[] = [
+  'Лучше быть последним — первым, чем первым — последним.',
+  'На случай, если буду нужен, то я там же, где и был, когда был не нужен.',
+  'Если волк молчит то лучше его не перебивай.',
+  'Каждый в цирке думает, что знает в цирке, но не каждый, что в цирке знает, что в цирке не каждый знает думает.',
+  'Легко вставать, когда ты не ложился.',
+  'За двумя зайцами погонишься — рыбку из пруда не выловишь, делу время, а отмеришь семь раз.',
+  'Каждый думает, что не знает что, но не каждый не знает, что знает, кто не я...',
+  'Кем бы ты ни был, кем бы ты не стал, помни, где ты был и кем ты стал.',
+];
+
+export const defaultMessages = [
+  {
+    id: '1',
+    author: 'Jacque Fresco',
+    text: 'Welcome to the chat',
+  },
+];
