+++ src/index.css
@@ -9,10 +9,3 @@ body {
 code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New", monospace;
 }
-
-.app-wrapper {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  align-items: stretch;
-}
