+++ src/utils/arrRandom.tsx
@@ -1,0 +1,4 @@
+export function arrRandom(arr: string[]) {
+  const rand = Math.floor(Math.random() * arr.length);
+  return arr[rand];
+}
