+++ tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+          "plugins": [{ "name": "typescript-plugin-css-modules" }],
     /* Visit https://aka.ms/tsconfig.json to read more about this file */
 
     /* Projects */
@@ -98,5 +99,6 @@
     // "skipDefaultLibCheck": true,                      /* Skip type checking .d.ts files that are included with TypeScript. */
     "skipLibCheck": true /* Skip type checking all .d.ts files. */
   },
-  "include": ["src/**/*"]
+  "include": ["src/**/*"],
+  "files": ["./types/types.d.ts"]
 }
