+++ types/types.d.ts
@@ -1,0 +1,2 @@
+declare module '*.module.css';
+declare module '*.module.scss';
